Consolidate completion signals and extract BuildSlotMap helper
- Merge 7 completion signal files into single CompletionSignals.cs
  (ICompletionSignal, BreakCompletionSignal, ContinueCompletionSignal,
  ReturnCompletionSignal, ThrowFlowCompletionSignal, YieldCompletionSignal,
  PendingAwaitCompletionSignal)
- Extract BuildSlotMap helper to EmitContext.cs to eliminate duplication
  between LoopEmitter and ForOfEmitter
