--- src/Asynkron.JsEngine/ContinueCompletionSignal.cs
@@ -1,12 +1,0 @@
-#region
-
-using Asynkron.JsEngine.Ast;
-
-#endregion
-
-namespace Asynkron.JsEngine;
-
-/// <summary>
-///     Signal indicating a continue statement was encountered.
-/// </summary>
-internal sealed record ContinueCompletionSignal(Symbol? Label = null) : ICompletionSignal;
