--- src/Asynkron.JsEngine/ICompletionSignal.cs
@@ -1,9 +1,0 @@
-namespace Asynkron.JsEngine;
-
-/// <summary>
-///     Represents a control flow signal used to manage JavaScript control flow statements
-///     (return, break, continue, yield, throw, pending await) as typed result values instead of state machine.
-/// </summary>
-public interface ICompletionSignal
-{
-}
