--- src/Asynkron.JsEngine/ReturnCompletionSignal.cs
@@ -1,15 +1,0 @@
-#region
-
-using Asynkron.JsEngine.JsTypes;
-
-#endregion
-
-namespace Asynkron.JsEngine;
-
-/// <summary>
-///     Signal indicating a return statement was encountered.
-/// </summary>
-internal sealed class ReturnCompletionSignal(JsValue jsValue) : ICompletionSignal
-{
-    public JsValue JsValue { get; } = jsValue;
-}
