--- src/Asynkron.JsEngine/ThrowFlowCompletionSignal.cs
@@ -1,15 +1,0 @@
-#region
-
-using Asynkron.JsEngine.JsTypes;
-
-#endregion
-
-namespace Asynkron.JsEngine;
-
-/// <summary>
-///     Signal indicating a throw statement was encountered.
-/// </summary>
-internal sealed class ThrowFlowCompletionSignal(JsValue jsValue) : ICompletionSignal
-{
-    public JsValue JsValue { get; } = jsValue;
-}
